+++ tests/querytree/src/testQueryTree.cpp
@@ -592,8 +592,8 @@ int main( int argc, const char* argv[])
 
 		int nofTests = parseInt( argv[2]);
 		int maxTreeSize = parseInt( argv[3]);
-		int typemax = (int)(strus::Math::sqrt(maxTreeSize+1));
-		int opmax = (int)(strus::Math::log(maxTreeSize+1));
+		int typemax = (int)(strus::Math::sqrt( (double)maxTreeSize+1));
+		int opmax = (int)(strus::Math::log( (double)maxTreeSize+1));
 		int valuemax = maxTreeSize+1;
 		static const char* desttypes[] = { "even", "odd", 0 };
